NL-221 Clear the warnings -Xlint:all reports
Drops a redundant cast, seals the two never-subclassed servlet classes
that let `this` escape their constructors, suppresses the two removal
warnings the Servlet 6.0 interface forces, and removes the try-with-
resources shapes and the Cookie.setVersion no-op test javac flags.

+++ netty-loom-spring-mvc/src/main/java/io/github/azholdaspaev/nettyloomspring/mvc/servlet/NettySessionManager.java
@@ -119,7 +119,7 @@ public class NettySessionManager {
             setDefaultMaxInactiveInterval(0);
             return;
         }
-        setDefaultMaxInactiveInterval((int) Math.clamp(timeout.toSeconds(), 1, Integer.MAX_VALUE));
+        setDefaultMaxInactiveInterval(Math.clamp(timeout.toSeconds(), 1, Integer.MAX_VALUE));
     }
 
     public NettySessionCookieConfig getCookieConfig() {
